+++ benchmark/benchmark.cpp
@@ -451,3 +451,25 @@ static void in_range(Function f, T lower_bound, T upper_bound,
 
 BENCH_IN_RANGE(in_range_u8, in_range_u8, 1, 2, uint8_t);
 BENCH_IN_RANGE(in_range_f32, in_range_f32, 1.111, 1.112, float);
+
+static void blur_and_downsample_u8(benchmark::State& state) {
+  kleidicv_filter_context_t* context;
+  kleidicv_error_t err = kleidicv_filter_context_create(
+      &context, 1, 5, 5, image_width, image_height);
+  if (err != KLEIDICV_OK) {
+    state.SkipWithError(
+        "Could not initialize filter context for Blur and Downsample");
+    return;
+  }
+
+  bench_functor(state, [context]() {
+    (void)kleidicv_blur_and_downsample_u8(
+        get_source_buffer_a<uint8_t>(), image_width * sizeof(uint8_t),
+        image_width, image_height, get_destination_buffer<uint8_t>(),
+        ((image_width + 1) / 2) * sizeof(uint8_t), 1,
+        KLEIDICV_BORDER_TYPE_REFLECT, context);
+  });
+
+  (void)kleidicv_filter_context_release(context);
+}
+BENCHMARK(blur_and_downsample_u8);