+++ scripts/ci-opencv.sh
@@ -49,6 +49,7 @@ IMGPROC_TEST_PATTERNS=(
   '*Imgproc_Resize*'
   '*Imgproc_Dilate*'
   '*Imgproc_Erode*'
+  '*Imgproc_PyramidDown*'
 )
 IMGPROC_TEST_PATTERNS_STR="$(join_strings_with_colon "${IMGPROC_TEST_PATTERNS[*]}")"
 ../../../conformity/opencv_kleidicv/bin/opencv_test_imgproc \