+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This changelog aims to follow the guiding principles of
 - Remap implementation
   - 2-channel s16 and s16+u16 fixed-point coordinates and 1-channel u8 input.
   - 2-channel s16 + 5+5 bits' fractions fixed-point coordinates and 1-channel u8 input.
+- Implementation for cv::pyrDown in the OpenCV HAL.
 
 ## 0.2.0 - 2024-09-30